functions/requirements.txt
annotated-types==0.7.0
argcomplete==1.10.3
beautifulsoup4==4.8.2
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.9.0
compressed_rtf==1.0.6
cryptography==42.0.7
deepdiff==7.0.1
deprecation==2.1.0
docstring_parser==0.16
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
ffmpeg-python==0.2.0
firebase-admin==6.5.0
firebase-functions==0.4.1
Flask==3.0.3
Flask-Cors==4.0.1
functions-framework==3.7.0
future==1.0.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.56.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-logging==3.10.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-events==0.12.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
IMAPClient==2.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
m3u8==5.0.0
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.0
olefile==0.47
opencc-python-reimplemented==0.1.7
ordered-set==4.1.0
packaging==24.0
pdfminer.six==20191110
pillow==10.3.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.2.1
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
shapely==2.0.4
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
textract==1.6.5
typesense==0.21.0
typing_extensions==4.12.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
xlrd==1.2.0
XlsxWriter==3.2.0
